Split graphql schema, typeDefs and resolvers into separate files
